+++ src/Controllers/DefaultController.php
@@ -3,24 +3,14 @@ namespace Application\Controllers;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Application\App\Http\RedirectResponseHttp;
+use Application\Application\Templating\TwigTrait;
 
 class DefaultController
 {
+    use TwigTrait;
+
     public function getHomePage (ServerRequestInterface $request){
 
-        $response = new RedirectResponseHttp('/accueil');
-        $response->send();
-
     }
-
-
-
-
-
-
-
-
-
-
     
 }
